refactor : 에러핸들러 수정
