+++ src/shared/api/socketService.ts
@@ -1,10 +1,4 @@
-import type {
-  ChatMessage,
-  SendMessage,
-  UpdateMessage,
-  WebRTCSignalData,
-  WebSocketMessage,
-} from '../model/types/webSocket';
+import type { WebSocketMessage } from '../model/types/webSocket';
 import { useRTCStore } from '../model/store/RTCStore';
 import { useAuthStore } from '../model/store/authStore';
 import { useChatStore } from '../model/store/chatStore';
@@ -13,29 +7,12 @@ import { useUnreadMessagesStore } from '../model/store/unreadMessagesStore';
 import { ROUTES } from '../model/constants/routes';
 import { queryClient } from './queryClient';
 import { QUERY_KEYS } from './queryKeys';
-
-const isDMMessage = (data: WebSocketMessage['data']): data is ChatMessage | SendMessage => {
-  return 'userId' in data && 'recipientId' in data;
-};
-
-const isSendMessage = (data: WebSocketMessage['data']): data is SendMessage => {
-  return 'content' in data && 'createdAt' in data && 'name' in data;
-};
-
-const isUpdateMessage = (data: WebSocketMessage['data']): data is UpdateMessage => {
-  return 'chatId' in data && 'content' in data && !('createdAt' in data);
-};
-
-const isWebRTCSignalData = (data: WebSocketMessage['data']): data is WebRTCSignalData => {
-  return 'fromUserId' in data && 'toUserId' in data;
-};
+import { isDMMessage, isSendMessage, isUpdateMessage, isWebRTCSignalData } from './typeGuards';
 
 export const SocketService = {
   handleDM: (wsMessage: WebSocketMessage) => {
     const { operation, data } = wsMessage;
     const myUserId = useAuthStore.getState().user?.id;
-    // WebRTC 시그널링 메시지는 다른 핸들러로 처리
-    if (isWebRTCSignalData(data)) return;
 
     // DM 메시지가 아니면 처리하지 않음
     if (!isDMMessage(data)) return;
